+++ source/pythonpath/easydev/debug.py
@@ -15,11 +15,15 @@ from easydev.setting import (
     INFOBOX,
     NAME_EXT,
     OS,
+    PY2,
     TITLE_DEBUG,
     WIN,
     WRITER,
 )
 
+
+if PY2:
+    str = unicode
 
 log = logging.getLogger(NAME_EXT)
 
@@ -75,11 +79,9 @@ class Debug(XDebug, LODefault):
         return mb.execute()
 
     def mri(self, obj):
-        m = self._create_instance('mytools.Mri')
+        m = comun.mri(obj)
         if m is None:
             self.msgbox('La extensión MRI no esta instalada', ERRORBOX)
-            return
-        m.inspect(obj)
         return
 
     def test(self, args):