+++ source/pythonpath/easydev/localc.py
@@ -4,8 +4,11 @@ import logging
 from easydev import comun
 from easydev.loapp import LOApp
 from org.universolibre.EasyDev import XLOCalc
-from easydev.setting import LOG, NAME_EXT
+from easydev.setting import LOG, NAME_EXT, PY2, SRV_GOS
 
+
+if PY2:
+    str = unicode
 
 log = logging.getLogger(NAME_EXT)
 
@@ -42,6 +45,13 @@ class LOCalc(XLOCalc, LOApp):
             index = doc.getSheets().getCount() + index
         return doc.getSheets().getByIndex(index)
 
+    def sheetActivate(self, address):
+        doc = self._get_doc(address.Doc)
+        address.Doc = doc
+        sheet = self.getSheet(address)
+        doc.getCurrentController().setActiveSheet(sheet)
+        return sheet
+
     def sheetInsert(self, address, pos, rename):
         doc = self._get_doc(address.Doc)
         index = pos
@@ -58,6 +68,28 @@ class LOCalc(XLOCalc, LOApp):
                     name = self._get_new_name(doc, name)
                 doc.getSheets().insertNewByName(name, index)
         return
+
+    def sheetInsertFromDoc(self, source, target, pos, values):
+        try:
+            src_doc = self._get_doc(source.Doc)
+            doc = target.Doc
+            index = pos
+            if pos < 0:
+                index = doc.getSheets().getCount() + pos + 1
+            if source.Sheet:
+                names = self._get_sheets_names(source.Sheet, src_doc, False)
+            else:
+                names = self.getSheetsNames(src_doc)
+            sheet = None
+            for name in names:
+                new_pos = doc.getSheets().importSheet(src_doc, name, index)
+                sheet = doc.getSheets().getByIndex(new_pos)
+                if values:
+                    pass
+            doc.getCurrentController().setActiveSheet(sheet)
+        except:
+            log.error('Insert from doc', exc_info=True)
+        return sheet
 
     def _get_sheet_name(self, name, doc):
         if isinstance(name, int):
@@ -170,7 +202,15 @@ class LOCalc(XLOCalc, LOApp):
             sheet = doc.getSheets().getByName(new_name)
             sheet.link(src_doc.URL, name, '', '', mode)
             sheet.setLinkMode(0)
-        print ('ok')
+            doc.getCurrentController().setActiveSheet(sheet)
+            dp = sheet.getDrawPage()
+            src_dp = src_doc.getSheets().getByName(name).getDrawPage()
+            for i in range(src_dp.getCount()):
+                src_img = src_dp.getByIndex(i)
+                src_doc.getCurrentController().select(src_img)
+                comun.copy(src_doc)
+                new = comun.paste(doc)
+                new.getByIndex(0).setPosition(src_img.getPosition())
         return
 
     def getCell(self, address):
@@ -205,6 +245,46 @@ class LOCalc(XLOCalc, LOApp):
             rango = sheet.getCellRangeByPosition(
                 address.Col, address.Row, address.EndCol, address.EndRow)
         return rango
+
+    def addRanges(self, container, ranges):
+        if container.ImplementationName != 'ScCellRangesObj':
+            container = container.createInstance('com.sun.star.sheet.SheetCellRanges')
+        if not isinstance(ranges, tuple):
+            ranges = (ranges,)
+        for r in ranges:
+            if not r:
+                continue
+            if not container.hasByName(r.AbsoluteName):
+                container.addRangeAddress(r.getRangeAddress(), False)
+        return container
+
+    def removeRanges(self, container, ranges):
+        if not isinstance(ranges, tuple):
+            ranges = (ranges,)
+        for r in ranges:
+            if container.hasByName(r.AbsoluteName):
+                container.removeRangeAddress(r.getRangeAddress())
+        return container
+
+    def getColumns(self, doc, ranges):
+        container = doc.createInstance('com.sun.star.sheet.SheetCellRanges')
+        if not isinstance(ranges, tuple):
+            ranges = (ranges,)
+        for r in ranges:
+            for c in range(r.getColumns().getCount()):
+                col = r.getColumns().getByIndex(c)
+                container.addRangeAddress(col.getRangeAddress(), False)
+        return container
+
+    def getRows(self, doc, ranges):
+        container = doc.createInstance('com.sun.star.sheet.SheetCellRanges')
+        if not isinstance(ranges, tuple):
+            ranges = (ranges,)
+        for r in ranges:
+            for r2 in range(r.getRows().getCount()):
+                row = r.getRows().getByIndex(r2)
+                container.addRangeAddress(row.getRangeAddress(), False)
+        return container
 
     def selectRange(self, doc, rango):
         doc.getCurrentController().select(rango)
@@ -280,4 +360,5 @@ class LOCalc(XLOCalc, LOApp):
         if comun.is_cell(cell):
             rango = self.getCurrentRegion(cell, False)
         return rango.queryVisibleCells()
+