+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+v 2.3.0 [27-Nov-2015]
+---------------------
+    - [Feature] Create containers ranges, add and remove ranges.
+    - [Feature] Get columns or rows from cell range.
+    - [Feature] Add copy, paste functions.
+    - [Feature] Insert sheets from documents, only LibreOffice.
+    - [Feature] Add properties for series
+    - [Feature] Add images and shapes
+    - [Feature] Zip, unzip files and folders
+    - [Fix] Unicode in OpenOffice PY2
+    - [Fix] Method getSizeScreen, only Windows
+
 v 2.2.0 [15-Nov-2015]
 ---------------------
     - [Feature] Get sheets names